+++ face_cropping/dnn_training/train_detector.py
@@ -21,7 +21,7 @@ def main():
                         help='Choose the dataset type', required=True)
     parser.add_argument('--output_path', type=str, help='Choose the output path', required=True)
 
-    parser.add_argument('--channel_scale', type=int, help='Choose the channel scale', required=True)
+    parser.add_argument('--channel_scale', type=float, help='Choose the channel scale', required=True)
     parser.add_argument('--head_kernel_size', type=int, help='Choose the head kernel size', required=True)
     parser.add_argument('--activation', choices=['relu', 'silu'], help='Choose the activation', required=True)
     parser.add_argument('--image_size', type=int, help='Choose the image width and height', required=True)
